requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
edge-tts==7.2.7
fastapi==0.128.5
Flask==3.1.2
Flask-SocketIO==5.6.0
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.32.5
rich==14.3.2
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.52.1
tabulate==0.9.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.5
wsproto==1.3.2
yarl==1.22.0
